BlinkingBillboard should not use long double for storing the time because of a precision loss on certain configurations. Instead, wrap around before computing the sine.

+++ code/branches/kicklib2/src/orxonox/graphics/BlinkingBillboard.h
@@ -74,7 +74,7 @@ namespace orxonox
             float frequency_;
             Degree phase_;
             bool bQuadratic_;
-            long double time_;
+            float time_;
     };
 }
